BACKEND/requirements.txt
flask==3.0.2
google-generativeai==0.3.1
requests==2.31.0
twilio==9.0.0
